--- src/main/java/Duke.java
@@ -1,50 +1,0 @@
-/**
- * A chatbot based on Project Duke
- *
- * @author KelvinSoo
- * @version A-MoreOOP
- */
-public class Duke {
-
-    private final Storage storage;
-    private TaskList taskList;
-    private final Ui ui;
-
-    /**
-     * A constructor to initialize a new chatbot.
-     *
-     * @param filePath Filepath of a saved chatbot or to save a chatbot
-     */
-    public Duke(String filePath) {
-        ui = new Ui();
-        storage = new Storage(filePath);
-        try {
-            taskList = new TaskList(storage.load());
-        } catch (DukeException e) {
-            ui.showError(e.getMessage());
-            taskList = new TaskList();
-        }
-    }
-
-    /**
-     * Start a new chatbot session
-     */
-    private void run() {
-        ui.greetUser();
-        boolean isExit = false;
-        while (!isExit) {
-            try {
-                String fullCommand = ui.readCommand();
-                Command c = Parser.parse(fullCommand);
-                c.execute(taskList, ui, storage);
-                isExit = c.isExit();
-            } catch (DukeException e) {
-                ui.showError(e.getMessage());
-            }
-        }
-    }
-
-    public static void main(String[] args) {
-        new Duke("./data/duke.txt").run();
-    }
-}
