+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,15 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.ui.Ui;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.ToDo;
+import duke.task.Event;
+import duke.utility.Utility;
+
 import java.time.LocalDateTime;
 
 /**
