+++ src/main/java/duke/utility/Utility.java
@@ -1,3 +1,5 @@
+package duke.utility;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
